+++ packages/admin/cms-admin/src/generator/future/generateForm.ts
@@ -1,8 +1,9 @@
 import { IntrospectionQuery } from "graphql";
 
+import { generateFormLayout } from "./generateForm/generateFormLayout";
 import { getForwardedGqlArgs } from "./generateForm/getForwardedGqlArgs";
 import { generateFormField } from "./generateFormField";
-import { FormConfig, FormFieldConfig, GeneratorReturn } from "./generator";
+import { FormConfig, FormFieldConfig, GeneratorReturn, isFormFieldConfig, isFormLayoutConfig } from "./generator";
 import { findMutationTypeOrThrow } from "./utils/findMutationType";
 import { generateImportsCode, Imports } from "./utils/generateImportsCode";
 import { isFieldOptional } from "./utils/isFieldOptional";
@@ -40,6 +41,16 @@ export function generateForm(
 
     const createMutationType = addMode && findMutationTypeOrThrow(config.createMutation ?? `create${gqlType}`, gqlIntrospection);
 
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    const formFields = config.fields.reduce<FormFieldConfig<any>[]>((acc, field) => {
+        if (isFormLayoutConfig(field)) {
+            acc.push(...field.fields);
+        } else if (isFormFieldConfig(field)) {
+            acc.push(field);
+        }
+        return acc;
+    }, []);
+
     const gqlArgs: ReturnType<typeof getForwardedGqlArgs>["gqlArgs"] = [];
     if (createMutationType) {
         const {
@@ -47,7 +58,7 @@ export function generateForm(
             props: forwardedGqlArgsProps,
             gqlArgs: forwardedGqlArgs,
         } = getForwardedGqlArgs({
-            fields: config.fields,
+            fields: formFields,
             gqlOperation: createMutationType,
             gqlIntrospection,
         });
@@ -66,7 +77,7 @@ export function generateForm(
 
     const { formPropsTypeCode, formPropsParamsCode } = generateFormPropsCode(props);
 
-    const rootBlockFields = config.fields
+    const rootBlockFields = formFields
         .filter((field) => field.type == "block")
         .map((field) => {
             // map is for ts to infer block type correctly
@@ -80,10 +91,10 @@ export function generateForm(
         });
     });
 
-    const numberFields = config.fields.filter((field) => field.type == "number");
-    const booleanFields = config.fields.filter((field) => field.type == "boolean");
-    const dateFields = config.fields.filter((field) => field.type == "date");
-    const readOnlyFields = config.fields.filter((field) => field.readOnly);
+    const numberFields = formFields.filter((field) => field.type == "number");
+    const booleanFields = formFields.filter((field) => field.type == "boolean");
+    const dateFields = formFields.filter((field) => field.type == "date");
+    const readOnlyFields = formFields.filter((field) => field.readOnly);
 
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const isOptional = (fieldConfig: FormFieldConfig<any>) => {
@@ -93,18 +104,33 @@ export function generateForm(
     let hooksCode = "";
     let formValueToGqlInputCode = "";
 
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    const generateFormFieldCode = (field: FormFieldConfig<any>) => {
+        const generated = generateFormField({ gqlIntrospection, baseOutputFilename }, field, config);
+        for (const name in generated.gqlDocuments) {
+            gqlDocuments[name] = generated.gqlDocuments[name];
+        }
+        imports.push(...generated.imports);
+        hooksCode += generated.hooksCode;
+        formValueToGqlInputCode += generated.formValueToGqlInputCode;
+        formFragmentFields.push(generated.formFragmentField);
+        return generated.code;
+    };
+
     const formFragmentFields: string[] = [];
     const fieldsCode = config.fields
         .map((field) => {
-            const generated = generateFormField({ gqlIntrospection, baseOutputFilename }, field, config);
-            for (const name in generated.gqlDocuments) {
-                gqlDocuments[name] = generated.gqlDocuments[name];
+            if (isFormFieldConfig(field)) {
+                return generateFormFieldCode(field);
+            } else if (isFormLayoutConfig(field)) {
+                const formLayoutFieldsCode = field.fields.map((field) => generateFormFieldCode(field)).join("\n");
+                const generated = generateFormLayout(field, formLayoutFieldsCode, config);
+                imports.push(...generated.imports);
+                formFragmentFields.push(...generated.formFragmentFields);
+                return generated.code;
+            } else {
+                throw new Error("Not supported config");
             }
-            imports.push(...generated.imports);
-            hooksCode += generated.hooksCode;
-            formValueToGqlInputCode += generated.formValueToGqlInputCode;
-            formFragmentFields.push(generated.formFragmentField);
-            return generated.code;
         })
         .join("\n");