+++ packages/site/cms-site/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @comet/cms-site
 
+## 7.0.0-beta.6
+
 ## 7.0.0-beta.5
 
 ### Major Changes
@@ -221,6 +223,8 @@
     Requires following changes to admin
 
     -   The `SitesConfig` must provide a `sitePreviewApiUrl`
+
+## 6.16.0
 
 ## 6.15.1